- security bugfix: plain text (and any) passwords aren't stored in session - little improvement: don't query database for user's data if user is already logged in (retrive it from session) - security bugfix: excluded 'sessions' table from database backup

+++ lib/LMS.class.php
@@ -144,6 +144,10 @@ class LMS
 		{
 			foreach($this->DB->ListTables() as $tablename)
 			{
+				// skip sessions table for security 
+				if($tablename == 'sessions')
+					continue;
+					
 				fputs($dumpfile,"DELETE FROM $tablename;\n");
 				$this->DB->Execute('SELECT * FROM '.$tablename);
 				while($row = $this->DB->_driver_fetchrow_assoc())
